+++ Backend/microservices/connector/src/main/java/com/ms/connector/dto/response/RestApiResponse.java
@@ -1,7 +1,12 @@
-package com.ms.connector.dto.apiresponse;
+package com.ms.connector.dto.response;
+
+import com.ms.connector.dto.response.data.RestApiResponseData;
+import lombok.Getter;
+import lombok.Setter;
 
 import java.util.List;
 
+@Setter
 public class RestApiResponse extends ApiResponse {
 
     private List<RestApiResponseData> data;
