+++ Backend/microservices/connector/src/main/java/com/ms/connector/dto/response/SoapApiResponse.java
@@ -1,7 +1,11 @@
-package com.ms.connector.dto.apiresponse;
+package com.ms.connector.dto.response;
+
+import com.ms.connector.dto.response.data.SoapApiResponseData;
+import lombok.Setter;
 
 import java.util.List;
 
+@Setter
 public class SoapApiResponse extends ApiResponse {
 
     private List<SoapApiResponseData> data;
