+++ Backend/microservices/connector/src/main/java/com/ms/connector/dto/response/WebsocketApiResponse.java
@@ -1,7 +1,11 @@
-package com.ms.connector.dto.apiresponse;
+package com.ms.connector.dto.response;
+
+import com.ms.connector.dto.response.data.WebsocketResponseData;
+import lombok.Setter;
 
 import java.util.List;
 
+@Setter
 public class WebsocketApiResponse extends ApiResponse {
 
     private List<WebsocketResponseData> data;
