+++ Backend/microservices/connector/src/main/java/com/ms/connector/dto/response/data/WebsocketResponseData.java
@@ -1,4 +1,4 @@
-package com.ms.connector.dto.apiresponse;
+package com.ms.connector.dto.response.data;
 
 public class WebsocketResponseData extends ApiResponseData {
