+++ Backend/microservices/connector/src/main/java/com/ms/connector/service/api/WebsocketApiConnection.java
@@ -1,7 +1,7 @@
 package com.ms.connector.service.api;
 
 import com.customstarter.model.request.SearchRequestPayload;
-import com.ms.connector.dto.apiresponse.WebsocketApiResponse;
+import com.ms.connector.dto.response.WebsocketApiResponse;
 import com.ms.connector.service.api.converter.JsonBodyConverter;
 import com.ms.connector.service.client.WebsocketClient;
